moved tests to __tests__ subfolder per best practices
