+++ src/frontend/efiling-frontend/src/domain/package/package-review/__tests__/PackageReviewService.test.js
@@ -5,8 +5,8 @@ import {
   downloadSubmissionSheet,
   downloadSubmittedDocument,
   withdrawSubmittedDocument,
-} from "./PackageReviewService";
-import { getCourtData } from "../../../modules/test-data/courtTestData";
+} from "../PackageReviewService";
+import { getCourtData } from "../../../../modules/test-data/courtTestData";
 
 jest.mock("axios");
 FileSaver.saveAs = jest.fn();
